security: to remove owned games, a token should be sent within the request, increasing security

+++ backend/app/Http/Controllers/OwnedGameController.php
@@ -60,11 +60,7 @@ class OwnedGameController extends Controller
 
     //adiciona o jogo a lista de owned e retorna o novo conjunto do owned games
     public function addOwned(Request $request){
-        Log::info('teste1');
-
         try {
-            Log::info('teste2');
-
             $token = $request->bearerToken();
             $user_id = $request->input('user_id');
             $game_api_id = $request->input('game_api_id');
@@ -115,30 +111,49 @@ class OwnedGameController extends Controller
     //deleta o jodo da lista de owned e retorna o novo conjunto do owned games
     public function removeOwned (Request $request){
         try{
-        $user_id = $request->input('user_id');
-        $game_api_id = $request->input('game_api_id');
-        $game_api_ids = $request->input('game_api_ids');
-        OwnedGame::where('user_id', $user_id)
-        ->where('game_api_id', $game_api_id)
-        ->delete();
-        $remove_favorite = FavoritedGame::where('user_id', $user_id)
-        ->where('game_api_id', $game_api_id)
-        ->first();
-        //caso exista o jogo nos favoritos, deleta ele de lá
-        if($remove_favorite){
-            $remove_favorite->delete();
-        }
-        $owned_games = $this->checkOwnedGames($user_id, $game_api_ids);
-        $favorite_games = $this->checkFavoriteGames($user_id, $game_api_ids);
-    $response = [
-        'owned_games' => $owned_games,
-        'favorite_games' => $favorite_games
-    ];
-    return response()->json($response);
+            $token = $request->bearerToken();
+            $user_id = $request->input('user_id');
+            $game_api_id = $request->input('game_api_id');
+            $game_api_ids = $request->input('game_api_ids');
 
-    }catch (\Exception $e) {
-        return response()->json(['Erro ao deletar jogo' => $e->getMessage()], 500);
-    }
+            //requisição nao enviou token junto
+            if (!$token) {
+                return response()->json(['error' => 'Unauthorized'], 401);
+            }
+
+            //id do token na database e pegamos o token lá
+            $tokenId = explode('|', $token)[0];
+            $personalAccessToken = PersonalAccessToken::where('id', $tokenId)->first();
+            if ($personalAccessToken) {
+                $token_value = explode('|', $token)[1];
+                if (hash_equals($personalAccessToken->token, hash('sha256', $token_value))) {
+                    Log::info("Entrou no remove owned autenticado!");
+                    OwnedGame::where('user_id', $user_id)
+                    ->where('game_api_id', $game_api_id)
+                    ->delete();
+                    $remove_favorite = FavoritedGame::where('user_id', $user_id)
+                    ->where('game_api_id', $game_api_id)
+                    ->first();
+                    //caso exista o jogo nos favoritos, deleta ele de lá
+                    if($remove_favorite){
+                        $remove_favorite->delete();
+                    }
+                    $owned_games = $this->checkOwnedGames($user_id, $game_api_ids);
+                    $favorite_games = $this->checkFavoriteGames($user_id, $game_api_ids);
+                    $response = [
+                        'owned_games' => $owned_games,
+                        'favorite_games' => $favorite_games
+                    ];
+                    return response()->json($response);
+                } else{
+                     return response()->json(['error' => 'Unauthorized'], 401);
+                }
+            }else{
+                return response()->json(['error' => 'Unauthorized'], 401);
+            }
+        }catch (\Exception $e) {
+            return response()->json(['Erro ao deletar jogo' => $e->getMessage()], 500);
+        }
     }
 
     //specific game page
